LayoutTests/fast/scrolling/scroll-animator-overlay-scrollbars-hovered-expected.txt

CONSOLE MESSAGE: MainFrameView: mouseEnteredContentArea
CONSOLE MESSAGE: MainFrameView: mouseMovedInContentArea
CONSOLE MESSAGE: FrameView: mouseEnteredHorizontalScrollbar
CONSOLE MESSAGE: FrameView: mouseEnteredContentArea
CONSOLE MESSAGE: MainFrameView: mouseMovedInContentArea
CONSOLE MESSAGE: FrameView: mouseExitedHorizontalScrollbar
CONSOLE MESSAGE: FrameView: mouseExitedContentArea
CONSOLE MESSAGE: MainFrameView: mouseMovedInContentArea
PASS successfullyParsed is true

TEST COMPLETE
Test for https://bugs.webkit.org/show_bug.cgi?id=153304.
